2.03RC3 provisional
Feature improvements/changed behaviour:
- M584 formatting improvement when no extruders
- In CoreNG, increment I2C reset count when resetting the I2C system
- Recognise filament usage comment in Prusa slicer
- G53 is now available even when workplace coordinates not supported in build (cancels tool offsets)
- For E3D: scale feed rate in proportion to total mix, for serial extruder drives etc.
- Allow M203 max speeds lower than 1mm/sec
- Enable laser in Duet085 build
- M563 P# with no other parameters: better response formatting when no heaters or no drives
- Added extra diagnostics for when a filename is too long
Bug fixes:
- Problem with leadscrew move when driver numbers >= MaxAxes are used to drive Z motors
- Added missing newline at end of some error messages
- M585 L parameter was not working
- In resume.g a G1 R command goes to the wrong coordinates if workplace coordinate offsets are being used
- Homing files when workplace coordinate offsets were active cause other axes to move
- Resurrect.g all coordinates need to be machine coordinates
- Resurrect.g restore workplace # and workplace offsets
- Error with babystepping: incorrect sign when doing tool offset inverse transform, so it restores an incorrect Z position

+++ src/RepRapFirmware.h
@@ -107,7 +107,7 @@ typedef double floatc_t;					// type of matrix element used for calibration
 typedef float floatc_t;						// type of matrix element used for calibration
 #endif
 
-typedef uint32_t AxesBitmap;				// Type of a bitmap representing a set of axes
+typedef uint16_t AxesBitmap;				// Type of a bitmap representing a set of axes
 typedef uint32_t DriversBitmap;				// Type of a bitmap representing a set of driver numbers
 typedef uint32_t FansBitmap;				// Type of a bitmap representing a set of fan numbers
 typedef uint16_t Pwm_t;						// Type of a PWM value when we don't want to use floats
